# apps/api-admin/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/home/rules": {
      "get": {
        "operationId": "listRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Rules"]
      },
      "post": {
        "operationId": "createRule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Rules"]
      }
    },
    "/api/v1/home/rules/{id}": {
      "patch": {
        "operationId": "updateRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Rules"]
      },
      "delete": {
        "operationId": "deleteRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Rules"]
      }
    },
...
    "/api/v1/quests": {
      "get": {
        "operationId": "listQuests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Quests"]
      },
      "post": {
        "operationId": "createQuest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Quests"]
      }
    },
    "/api/v1/quests/{id}": {
      "patch": {
        "operationId": "updateQuest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Quests"]
      },
      "delete": {
        "operationId": "deleteQuest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Quests"]
      }
    },
...
    "/api/v1/users/{id}/quests/{questId}": {
      "patch": {
        "operationId": "setMemberQuestStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMemberQuestStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["Users"]
      }
    },
...
  },
  "info": {
    "title": "FanDrop Admin API",
    "description": "Admin backend: artist-scoped CRUD, moderation, uploads",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "stepNumber": {
            "type": "number"
          },
          "isVisible": {
            "type": "boolean"
          }
        },
        "required": ["id", "title", "body", "stepNumber", "isVisible"]
      },
      "CreateRuleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 30
          },
          "body": {
            "type": "string",
            "nullable": true,
            "maxLength": 90
          },
          "stepNumber": {
            "type": "number",
            "minimum": 0
          },
          "isVisible": {
            "type": "boolean",
            "default": true
          }
        },
        "required": ["title"]
      },
      "UpdateRuleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 30
          },
          "body": {
            "type": "string",
            "nullable": true,
            "maxLength": 90
          },
          "stepNumber": {
            "type": "number",
            "minimum": 0
          },
          "isVisible": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "QuestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "rewardPoints": {
            "type": "number"
          },
          "availableAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "featuredPos": {
            "type": "number"
          },
          "isVisible": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "link",
          "imageUrl",
          "rewardPoints",
          "availableAt",
          "featuredPos",
          "isVisible"
        ]
      },
      "CreateQuestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "link": {
            "type": "string",
            "maxLength": 2048
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "rewardPoints": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "availableAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "featuredPos": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "isVisible": {
            "type": "boolean",
            "default": true
          }
        },
        "required": ["title", "link"]
      },
      "UpdateQuestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "link": {
            "type": "string",
            "maxLength": 2048
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "rewardPoints": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "availableAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "featuredPos": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "isVisible": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "SetMemberQuestStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["NOT_STARTED", "COMPLETED"]
          }
        },
        "required": ["status"]
      }
    }
  }
}