# apps/web-admin/src/api/generated/model/createRuleDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * FanDrop Admin API
 * Admin backend: artist-scoped CRUD, moderation, uploads
 * OpenAPI spec version: 1.0
 */

export interface CreateRuleDto {
  /** @maxLength 30 */
  title: string;
  /**
     * @maxLength 90
     * @nullable
     */
  body?: string | null;
  /** @minimum 0 */
  stepNumber?: number;
  isVisible?: boolean;
}