# apps/web-admin/src/api/generated/model/questResponseDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * FanDrop Admin API
 * Admin backend: artist-scoped CRUD, moderation, uploads
 * OpenAPI spec version: 1.0
 */

export interface QuestResponseDto {
  id: string;
  title: string;
  /** @nullable */
  description: string | null;
  link: string;
  /** @nullable */
  imageUrl: string | null;
  rewardPoints: number;
  /** @nullable */
  availableAt: string | null;
  featuredPos: number;
  isVisible: boolean;
}